Changed all geometry from triangles to rectangles.
Updated data folder. I was able to make a fairly significant
optimization in the ray tracer by changing all pairs of triangles into
rectangles, since everything in the original game (sprites, walls,
textures) can be represented as a rectangle. This cut the geometry count
in half, and kernel performance appears to have improved by about 30%.
The UV coordinates are also unnecessary now for normal sized blocks
(1x1x1), but may be necessary for squished blocks (i.e., bed sides) that
only show a portion of the texture (unless the texture can be modified
to avoid this). Now my next focus will most likely be sprites.

+++ OpenTESArena/src/Rendering/TextureReference.cpp
@@ -1,6 +1,6 @@
 #include "TextureReference.h"
 
-TextureReference::TextureReference(int offset, int width, int height)
+TextureReference::TextureReference(int offset, short width, short height)
 {
 	this->offset = offset;
 	this->width = width;
@@ -17,12 +17,12 @@ int TextureReference::getOffset() const
 	return this->offset;
 }
 
-int TextureReference::getWidth() const
+short TextureReference::getWidth() const
 {
 	return this->width;
 }
 
-int TextureReference::getHeight() const
+short TextureReference::getHeight() const
 {
 	return this->height;
 }
